requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
requests>=2.32.0
pydantic-settings>=2.0.0
kubernetes>=29.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
tenacity>=9.0.0
